Move source files to src folder
