+++ bin/electron-packager.js
@@ -11,9 +11,9 @@ if (!semver.satisfies(process.versions.node, packageJSON.engines.node)) {
 }
 
 // Not consts so that this file can load in Node < 4.0
-var common = require('../common')
+var common = require('../src/common')
 var fs = require('fs')
-var packager = require('../')
+var packager = require('..')
 var path = require('path')
 var usage = fs.readFileSync(path.resolve(__dirname, '..', 'usage.txt')).toString()
