+++ src/win32.js
